+++ Pandora/Paradigm/Inventory.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE AllowAmbiguousTypes #-} {-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 
-module Pandora.Paradigm.Inventory (module Exports, zoom, intensify, magnify, (=<>), (~<>), adjust) where
+module Pandora.Paradigm.Inventory (module Exports, zoom, (=<>), (~<>)) where
 
 import Pandora.Paradigm.Inventory.Optics as Exports
 import Pandora.Paradigm.Inventory.Store as Exports
@@ -11,18 +12,15 @@ import Pandora.Paradigm.Inventory.Environment as Exports
 import Pandora.Paradigm.Inventory.Accumulator as Exports
 
 import Pandora.Pattern.Semigroupoid ((.))
-import Pandora.Pattern.Category (($), (#), identity)
+import Pandora.Pattern.Category (($))
 import Pandora.Pattern.Morphism.Flip (Flip (Flip))
 import Pandora.Pattern.Functor.Covariant ((<-|-))
 import Pandora.Pattern.Functor.Adjoint (Adjoint ((-|), (|-)))
-import Pandora.Pattern.Functor.Bivariant ((<->))
 import Pandora.Paradigm.Primary.Algebraic.Product ((:*:) ((:*:)))
 import Pandora.Paradigm.Primary.Algebraic.Exponential ((!.), (%))
 import Pandora.Paradigm.Primary.Algebraic (extract)
-import Pandora.Paradigm.Primary.Functor.Identity (Identity (Identity))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, (!))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
-import Pandora.Paradigm.Structure.Ability.Accessible (Accessible (access))
 
 instance Adjoint (->) (->) (Store s) (State s) where
 	(-|) :: (Store s a -> b) -> a -> State s b
@@ -38,14 +36,8 @@ instance Adjoint (->) (->) (Equipment e) (Environment e) where
 	f -| x = Environment $ f . Equipment -| x
 	g |- x = run . g |- run x
 
-zoom :: forall bg ls t a . Stateful bg t => Convex Lens bg ls -> State ls a -> t a
-zoom lens less = adapt . State $ (restruct |-) . run . run lens where
-
-	restruct :: (Identity ls -> bg) -> Identity ls -> bg :*: a
-	restruct to = (to . Identity <-> identity @(->)) . run less . extract @Identity
-
-intensify :: forall u bg ls t result . Stateful bg t => Lens u bg ls -> State (u ls) result -> t result
-intensify lens less = adapt . State $ \source -> restruct |- run (lens ! source) where
+zoom :: forall bg ls t u result . Stateful bg t => Lens u bg ls -> State (u ls) result -> t result
+zoom lens less = adapt . State $ \source -> restruct |- run (lens ! source) where
 
 	restruct :: (u ls -> bg) -> u ls -> bg :*: result
 	restruct to target = run $ to <-|- Flip (less ! target)
@@ -55,9 +47,3 @@ lens =<> new = modify $ set lens new
 
 (~<>) :: Stateful src t => Lens available src tgt -> (available tgt -> available tgt) -> t src
 lens ~<> f = modify $ over lens f
-
-magnify :: forall bg ls t . (Accessible ls bg, Stateful bg t) => t ls
-magnify = zoom @bg # access @ls @bg # current
-
-adjust :: forall bg ls t . (Accessible ls bg, Stateful bg t) => (ls -> ls) -> t ls
-adjust = zoom @bg (access @ls) . modify